+++ src/scicloj/metamorph/ml/classification.clj
@@ -103,4 +103,15 @@
       (ds/new-dataset [(ds/new-column target-column-name dummy-labels {:column-type :prediction
                                                                        :categorical-map (get target-categorical-maps target-column-name)})])))
      
-  {:glance-fn (fn [_] (ds/->dataset {:npar 0}))})
+  {:glance-fn (fn [_] (ds/->dataset {:npar 0}))
+   :options
+   [:map {:closed true}
+    [:dummy-strategy
+     {:optional true
+      :description "The strategy to use for the dummy classifier"}
+     [:enum :majority-class :fixed-class :random-class]]
+    [:fixed-class {:optional true
+                   :description "The fixed class to use when strategy is :fixed-class"}
+     number?]]})
+
+