+++ test/scicloj/metamorph/decl_pipe_test.clj
@@ -3,7 +3,6 @@
              [scicloj.metamorph.core :as mm]
              [scicloj.metamorph.ml :as ml]
              [scicloj.metamorph.ml.loss :as loss]
-             [tech.v3.dataset.metamorph :as ds-mm]
              [tablecloth.api :as tc]
              [clojure.string :as str]
              [tech.v3.dataset :as ds]))